requirements.txt
# Core
anthropic>=0.40.0
openai>=1.0.0
httpx>=0.27.0
click>=8.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-httpx>=0.30.0   # mock HTTP calls in tests

# Dev
ruff>=0.4.0            # linting + formatting
